+++ internal/client/addrproc_test.go
@@ -264,6 +264,100 @@ func TestDefaultAddrProc_Process_WHOIS(t *testing.T) {
 	}
 }
 
+func TestDefaultAddrProc_Process_WHOIS_PrivateSubnets(t *testing.T) {
+	t.Parallel()
+
+	wantInfo := &whois.Info{
+		City: testWHOISCity,
+	}
+
+	testCases := []struct {
+		privateSubnet netip.Prefix
+		ip            netip.Addr
+		wantInfo      *whois.Info
+		name          string
+		wantDial      bool
+	}{
+		{
+			privateSubnet: netip.MustParsePrefix("2606:4700:4700::/64"),
+			ip:            netip.MustParseAddr("2606:4700:4700::1111"),
+			name:          "private_network",
+			wantInfo:      nil,
+			wantDial:      false,
+		},
+		{
+			privateSubnet: netip.MustParsePrefix("2606:4700:4700::/64"),
+			ip:            netip.MustParseAddr("2001:4860:4860::8888"),
+			name:          "public",
+			wantInfo:      wantInfo,
+			wantDial:      true,
+		},
+		{
+			privateSubnet: netip.MustParsePrefix("192.168.0.0/16"),
+			ip:            netip.MustParseAddr("::ffff:192.168.0.1"),
+			name:          "private_network_mapped_ipv4",
+			wantInfo:      nil,
+			wantDial:      false,
+		},
+		{
+			privateSubnet: netip.MustParsePrefix("192.168.0.0/16"),
+			ip:            netip.MustParseAddr("::ffff:8.8.8.8"),
+			name:          "public_mapped_ipv4",
+			wantInfo:      wantInfo,
+			wantDial:      true,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+
+			var dialed bool
+			whoisConn := &fakenet.Conn{
+				OnClose: func() (err error) { return nil },
+				OnRead: func(b []byte) (n int, err error) {
+					data := "city: " + testWHOISCity + "\n"
+					copy(b, data)
+
+					return len(data), io.EOF
+				},
+				OnSetDeadline: func(_ time.Time) (err error) { return nil },
+				OnWrite:       func(b []byte) (n int, err error) { return len(b), nil },
+			}
+
+			updInfoCh := make(chan *whois.Info, 1)
+			p := client.NewDefaultAddrProc(&client.DefaultAddrProcConfig{
+				BaseLogger: slogutil.NewDiscardLogger(),
+				DialContext: func(_ context.Context, _, _ string) (conn net.Conn, err error) {
+					dialed = true
+
+					return whoisConn, nil
+				},
+				PrivateSubnets: tc.privateSubnet,
+				AddressUpdater: &aghtest.AddressUpdater{
+					OnUpdateAddress: func(
+						_ context.Context,
+						_ netip.Addr,
+						_ string,
+						info *whois.Info,
+					) {
+						updInfoCh <- info
+					},
+				},
+				UseWHOIS: true,
+			})
+			testutil.CleanupAndRequireSuccess(t, p.Close)
+
+			ctx := testutil.ContextWithTimeout(t, testTimeout)
+			p.Process(ctx, tc.ip)
+
+			gotInfo, _ := testutil.RequireReceive(t, updInfoCh, testTimeout)
+			assert.Equal(t, tc.wantInfo, gotInfo)
+			assert.Equal(t, tc.wantDial, dialed)
+		})
+	}
+}
+
 func TestDefaultAddrProc_Close(t *testing.T) {
 	t.Parallel()